added documentation; still debugging page sync issues.
problem: go to end of chapter 1, press play. let the audio play through
to chapter 2. notice the text highlighting has ceased although the
audio continues. the problem seems to be that the pagination view does
not hear the 'change:mo_text_id' events for chapter 2. has the
pagination view been created in time for the new section yet?