fixed MO/pagination sync
There were two main problems:
1. MO controller acted as a global controller, but the pagination view
expected a new instance with every new pagination view, which caused
conflicts in handling events after crossing spine boundaries. It has
been changed so that there is only one instance in all of Readium, and
this instance lives under EpubController.

2. When crossing spine boundaries for reflowable content, Readium
re-paginates the content and the page number changes a few times. MO
must wait for a callback from the PaginationViewBase to let it know
that the pages have been loaded. Otherwise, MO tries to start playing
and observes a page change, which it interprets to have come from the
user. This resulted in  MO skipping ahead a page when advancing to the
next spine item.